school-backend/requirements.txt
asgiref==3.5.0
Django==3.0.7
djangorestframework==3.13.1
Faker==13.4.0
importlib-metadata==4.11.3
Markdown==3.3.6
python-dateutil==2.8.2
python-dotenv==0.20.0
python-status==1.0.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
validate-docbr==1.10.0
zipp==3.8.0
